refactor(executor, runner): streamline session management and remove deep copy function
- Simplified session retrieval logic in the runner by consolidating session creation and retrieval into a single method.
- Removed the deepCopyAny and deepCopyState functions from the executor, as they were no longer needed.
- Updated comments for clarity and improved error handling in event processing.
- Cleaned up unnecessary code and improved readability in the runner's Run method.
