+++ runner/runner.go
@@ -112,27 +112,19 @@ func (r *runner) Run(
 	message model.Message,
 	runOpts ...agent.RunOption,
 ) (<-chan *event.Event, error) {
-
+	// Resolve or create the session for this user and conversation.
 	sessionKey := session.Key{
 		AppName:   r.appName,
 		UserID:    userID,
 		SessionID: sessionID,
 	}
 
-	// Get session or create if it doesn't exist.
-	sess, err := r.sessionService.GetSession(ctx, sessionKey)
+	sess, err := r.getOrCreateSession(ctx, sessionKey)
 	if err != nil {
 		return nil, err
 	}
-	if sess == nil {
-		if sess, err = r.sessionService.CreateSession(
-			ctx, sessionKey, session.StateMap{},
-		); err != nil {
-			return nil, err
-		}
-	}
 
-	// Create invocation.
+	// Build run options with defaults and construct the invocation.
 	ro := agent.RunOptions{RequestID: uuid.NewString()}
 	for _, opt := range runOpts {
 		opt(&ro)
@@ -180,7 +172,6 @@ func (r *runner) Run(
 		if err := r.sessionService.AppendEvent(ctx, sess, evt); err != nil {
 			return nil, err
 		}
-
 	}
 
 	// Ensure the invocation can be accessed by downstream components (e.g., tools)
@@ -195,13 +186,37 @@ func (r *runner) Run(
 		return nil, err
 	}
 
-	// Create a new channel for processed events.
+	// Process the agent events and emit them to the output channel.
+	return r.processAgentEvents(ctx, sess, invocation, agentEventCh), nil
+}
+
+// getOrCreateSession returns an existing session or creates a new one.
+func (r *runner) getOrCreateSession(
+	ctx context.Context, key session.Key,
+) (*session.Session, error) {
+	sess, err := r.sessionService.GetSession(ctx, key)
+	if err != nil {
+		return nil, err
+	}
+	if sess != nil {
+		return sess, nil
+	}
+	return r.sessionService.CreateSession(ctx, key, session.StateMap{})
+}
+
+// processAgentEvents consumes agent events, persists to session, and emits.
+func (r *runner) processAgentEvents(
+	ctx context.Context,
+	sess *session.Session,
+	invocation *agent.Invocation,
+	agentEventCh <-chan *event.Event,
+) chan *event.Event {
 	processedEventCh := make(chan *event.Event)
 	// Start a goroutine to process and append events to session.
 	go func() {
 		defer func() {
-			if r := recover(); r != nil {
-				log.Errorf("panic in runner event loop: %v\n%s", r, string(debug.Stack()))
+			if rr := recover(); rr != nil {
+				log.Errorf("panic in runner event loop: %v\n%s", rr, string(debug.Stack()))
 			}
 			close(processedEventCh)
 			invocation.CleanupNotice(ctx)
@@ -213,7 +228,8 @@ func (r *runner) Run(
 				continue
 			}
 			// Append event to session if it's complete (not partial).
-			if len(agentEvent.StateDelta) > 0 || (agentEvent.Response != nil && !agentEvent.IsPartial && agentEvent.IsValidContent()) {
+			if len(agentEvent.StateDelta) > 0 ||
+				(agentEvent.Response != nil && !agentEvent.IsPartial && agentEvent.IsValidContent()) {
 				if err := r.sessionService.AppendEvent(ctx, sess, agentEvent); err != nil {
 					log.Errorf("Failed to append event to session: %v", err)
 				}
@@ -234,7 +250,6 @@ func (r *runner) Run(
 				completionID := agent.GetAppendEventNoticeKey(agentEvent.ID)
 				invocation.NotifyCompletion(ctx, completionID)
 			}
-
 			if err := event.EmitEvent(ctx, processedEventCh, agentEvent); err != nil {
 				return
 			}
@@ -261,10 +276,10 @@ func (r *runner) Run(
 		// Send the runner completion event to output channel.
 		agent.EmitEvent(ctx, invocation, processedEventCh, runnerCompletionEvent)
 	}()
-
-	return processedEventCh, nil
+	return processedEventCh
 }
 
+// shouldAppendUserMessage checks if the incoming user message should be appended to the session.
 func shouldAppendUserMessage(message model.Message, seed []model.Message) bool {
 	if len(seed) == 0 {
 		return true
