+++ storage/elasticsearch/internal/elasticsearch/elasticsearch_v7_test.go
@@ -122,7 +122,7 @@ func TestClientV7_Count(t *testing.T) {
 			resp.Header.Set("X-Elastic-Product", "Elasticsearch")
 			return resp
 		}
-		
+
 		// POST /{index}/_count
 		if r.Method == http.MethodPost && strings.Contains(r.URL.Path, "_count") {
 			return ok(http.StatusOK, `{"count":3}`)
@@ -147,7 +147,7 @@ func TestClientV7_DeleteByQuery(t *testing.T) {
 			resp.Header.Set("X-Elastic-Product", "Elasticsearch")
 			return resp
 		}
-		
+
 		// POST /{index}/_delete_by_query
 		if r.Method == http.MethodPost && strings.Contains(r.URL.Path, "_delete_by_query") {
 			return ok(http.StatusOK, `{"deleted":2}`)
@@ -171,7 +171,7 @@ func TestClientV7_Refresh(t *testing.T) {
 			resp.Header.Set("X-Elastic-Product", "Elasticsearch")
 			return resp
 		}
-		
+
 		// POST /{index}/_refresh
 		if r.Method == http.MethodPost && strings.Contains(r.URL.Path, "_refresh") {
 			return ok(http.StatusOK, `{}`)
